edumind-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
psycopg[binary]==3.1.18
python-dotenv==1.0.1
